move unpackInputDataset to src/proc/steps
