+++ bin/check-input-datasets.js
@@ -3,11 +3,10 @@ const Ajv = require("ajv").default;
 
 const { DATA_FOLDER_NAME } = require("../src/process-single-dataset/constants");
 const {
-  readDatasetJson,
-  readAndParseFile,
-  readPossibleZippedFile,
+  readDatasetJson
 } = require("../src/utils");
 const dataPrep = require("../src/data-validation/data-prep");
+const unpackInputDataset = require("../src/process-single-dataset/steps/unpack-input-dataset")
 
 // referenced partial schemas
 const refSchemas = [
@@ -71,27 +70,6 @@ const checkForError = (fileName, json, schemaFileName) => {
   }
 };
 
-const unpackInputDataset = async (datasetReadFolder) => {
-  const datasetJson = await readDatasetJson(datasetReadFolder);
-  const featureDefs = await readAndParseFile(
-    `${datasetReadFolder}/${datasetJson.featureDefsPath}`
-  );
-  const images = await readAndParseFile(
-    `${datasetReadFolder}/${datasetJson.viewerSettingsPath}`
-  );
-  const measuredFeatures = await readPossibleZippedFile(
-    datasetReadFolder,
-    datasetJson.featuresDataPath
-  );
-
-  const inputDataset = {
-    dataset: datasetJson,
-    "feature-defs": featureDefs,
-    "measured-features": measuredFeatures,
-    images: images,
-  };
-  return inputDataset;
-};
 
 const validateDatasets = () => {
   fsPromises
